The service unit files need to be enabled to start on boot
To allow the service units file to start on boot, they need to be
enabled. If the role variable restic_rest_target_enable is set to 'true'
the service unit files get enabled.

I choose 'false' as default to not change the current behavior, but this
should be changed in the future.