Change to restic-server v0.10.0
 BREAKING CHANGES
-----------

I have not used the restic-server before. So I just changed what it needs to be compatible with the current version.

But how to keep it backwards compatible... I do not know. But I don't know if you want that either.

Here are the adjustments it needs for restic-server v0.10.0.