Add variable restic_rest_target_start
rest-server.target gets started when this role is deployed.

When restic_rest_target_enable is true, the services get enabled and
started on boot but the rest-server.target is not.
So when deploying this role after a reboot, rest-server.target gets
started again and causes ansible to report a change.

To prevent triggering a change we implemented the variable
`restic_rest_target_start` which prevents rest-server.target to be
started if set to false. While doing so we also make sure that the
restic service units are started anyway (if `restic_rest_target_enable` is
set to true). That way we do not need to start rest-server.target to
start the service units.

`restic_rest_target_start` defaults to true because we do not want to
change the default behavior, but it seems like we actually do not need
the target anymore.